chore(knockdog): 엔티티 social > social-user 변경, 파일 네이밍 및 구조 변경
