+++ apps/knockdog/src/entities/social-user/model/store/socialUserStore.ts
@@ -1,12 +1,12 @@
 import { create } from 'zustand';
 
-import { SocialUser } from './social-user';
+import { SocialUser } from '../socialUser';
 
 interface SocialUserStore {
   /** 소셜 유저 정보 */
   socialUser: SocialUser | null;
   /** 소셜 유저 정보 저장 */
-  setSocialUser: (socialUser: SocialUser) => void;
+  setSocialUser: (socialUser: SocialUser | null) => void;
 }
 
 export const useSocialUserStore = create<SocialUserStore>((set) => ({
