--- apps/knockdog/src/entities/social/index.ts
@@ -1,11 +1,0 @@
-export type {
-  SocialUser,
-  SocialProvider,
-  VerifyOidcResponse,
-  VerifyOidcTokenParams,
-  VerifyOidcResultCode,
-} from './model';
-
-export { SOCIAL_PROVIDER, VERIFY_OIDC_RESULT_CODE, SOCIAL_PROVIDER_KO, useSocialUserStore } from './model';
-
-export { postVerifyOidc } from './api';
